Android: Extracted the inner class `org.coolreader.db.CRDBService$Task` into a separate class `org.coolreader.db.Task`. Also, a base class has been created to implement any services: `org.coolreader.db.BaseService`. This is necessary to create other services with implementation similar to `org.coolreader.db.CRDBService` and reuse the code.
